OOTB Perf gains for new consumers
